Install script for database agents
In the context of Discover, we want to ease the installation process of
Teleport.

Our current focus is Databases.

Adding a database resource requires at least one Database service/agent
running with matching labels for that resource.

The focus of this PR is to ease the installation of those agents.

Those service/agents are currently installed in a multi step flow:
- install Teleport on the host (select OS/arch/distribution format)
- configure the service by running `teleport configure db ...`
- start the application (possibly configuring the systemd unit file.

As you can see, this is a lot of steps.

This PR adds a new installation script: install-database.sh
This script is actually the same used for installing Server Nodes or App
Service/Agents but now can also install database service/agents.

So, how does the new flow look like?
After adding the Database info (type, hostname, port and labels), and if
there's no agent available, a new step appears: add a database agent.
The user provides a list of labels they want the agent to monitor.
Those labels are stored as AgentMatcherLabels within the token which is
generate after confirming the labels.
The user then gets an URL which they can use to install the agent.
This URL contains the token and when invoked creates the
install-database.sh script containing the `db_service` configuration.
The field `db_service.resources.0.labels` contains the previously
defined labels.
This script is almost the same that exists today for installing Server
Nodes and App services/agents, it only changes the `teleport.yaml`
configuration.

When the user run this command, the UI gets notified and can then
proceed to the next step.